feat: require `read:session` to access `/user` endpoint
