+++ models/activation.js
@@ -114,7 +114,6 @@ async function activateUserByUserId(userId) {
   const activatedUser = await user.setFeatures(userId, [
     "create:session",
     "read:session",
-    "update:user",
   ]);
   return activatedUser;
 }
