+++ pages/api/v1/user/index.js
@@ -5,7 +5,8 @@ import { createRouter } from "next-connect";
 
 const router = createRouter();
 
-router.get(getHandler);
+router.use(controller.injectAnonymousOrUser);
+router.get(controller.canRequest("read:session"), getHandler);
 
 export default router.handler(controller.errorHandlers);
