--- tests/integration/api/v1/sessions/get.test.js
@@ -1,122 +1,0 @@
-import { version as uuidVersion } from "uuid";
-import orchestrator from "tests/orchestrator.js";
-import session from "models/session";
-import setCookieParser from "set-cookie-parser";
-
-beforeAll(async () => {
-  await orchestrator.waitForAllServices();
-  await orchestrator.clearDatabase();
-  await orchestrator.runPendingMigrations();
-});
-
-describe("GET /api/v1/sessions", () => {
-  describe("Default user", () => {
-    test("With valid session", async () => {
-      const createdUser = await orchestrator.createUser({
-        username: "UserWithValidSession",
-      });
-
-      const sessionObject = await orchestrator.createSession(createdUser.id);
-
-      const response = await fetch("http://localhost:3000/api/v1/user", {
-        headers: {
-          Cookie: `session_id=${sessionObject.token}`,
-        },
-      });
-
-      expect(response.status).toBe(200);
-
-      const responseBody = await response.json();
-
-      expect(responseBody).toEqual({
-        id: createdUser.id,
-        username: "UserWithValidSession",
-        email: createdUser.email,
-        password: createdUser.password,
-        features: ["read:activation_token"],
-        created_at: createdUser.created_at.toISOString(),
-        updated_at: createdUser.updated_at.toISOString(),
-      });
-
-      expect(uuidVersion(responseBody.id)).toBe(4);
-      expect(Date.parse(responseBody.created_at)).not.toBeNaN();
-      expect(Date.parse(responseBody.updated_at)).not.toBeNaN();
-
-      const renewedSessionObject = await session.findOneValidByToken(
-        sessionObject.token,
-      );
-      expect(
-        renewedSessionObject.expires_at > sessionObject.expires_at,
-      ).toEqual(true);
-      expect(
-        renewedSessionObject.updated_at > sessionObject.updated_at,
-      ).toEqual(true);
-
-      // Set-cookie assertions
-      const parsedSetCookie = setCookieParser(response, {
-        map: true,
-      });
-
-      expect(parsedSetCookie.session_id).toEqual({
-        name: "session_id",
-        value: sessionObject.token,
-        maxAge: session.EXPIRATION_IN_MILLISECONDS / 1000,
-        path: "/",
-        httpOnly: true,
-      });
-    });
-
-    test("With nonexistent session", async () => {
-      const nonexistentToken =
-        "f0b62a5ff97ae607701ceeee2e3c4987c4b9debb534410e2444f9eb2288b6e3b90158a71d086e31eabef9b36cbb549e1";
-
-      const response = await fetch("http://localhost:3000/api/v1/user", {
-        headers: {
-          cookie: `session_id=${nonexistentToken}`,
-        },
-      });
-
-      expect(response.status).toBe(401);
-
-      const responseBody = await response.json();
-
-      expect(responseBody).toEqual({
-        name: "UnauthorizedError",
-        message: "Usuário não possui sessão ativa.",
-        action: "Verifique se este usuário está logado e tente novamente.",
-        status_code: 401,
-      });
-    });
-
-    test("With expired session", async () => {
-      jest.useFakeTimers({
-        now: new Date(Date.now() - session.EXPIRATION_IN_MILLISECONDS),
-      });
-
-      const createdUser = await orchestrator.createUser({
-        username: "UserWithExpiredSession",
-      });
-
-      const sessionObject = await orchestrator.createSession(createdUser.id);
-
-      jest.useRealTimers();
-
-      const response = await fetch("http://localhost:3000/api/v1/user", {
-        headers: {
-          Cookie: `session_id=${sessionObject.token}`,
-        },
-      });
-
-      expect(response.status).toBe(401);
-
-      const responseBody = await response.json();
-
-      expect(responseBody).toEqual({
-        name: "UnauthorizedError",
-        message: "Usuário não possui sessão ativa.",
-        action: "Verifique se este usuário está logado e tente novamente.",
-        status_code: 401,
-      });
-    });
-  });
-});
